fix: implement adding pictures
